requirements.txt
google-genai
anthropic
modal
python-dotenv
rich
fastapi
uvicorn[standard]
python-dotenv
openai
